push kedua

+++ database/migrations/2019_12_05_125008_create_employees_table.php
@@ -15,7 +15,7 @@ class CreateEmployeesTable extends Migration
     {
         Schema::create('employees', function (Blueprint $table) {
             $table->increments('id_employees');
-            $table->integer('id_jobs')->unsigned()->nullable(); 
+            $table->bigInteger('id_jobs')->unsigned()->nullable(); 
             $table->string('name'); 
             $table->string('email'); 
             $table->string('phone');
